agentic/pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "docker>=7.1.0",
    "fastapi>=0.121.3",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "llama-index>=0.14.8",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-llms-ollama>=0.9.0",
    "llama-index-vector-stores-chroma>=0.5.3",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
